requirements.txt
# Core Dependencies for RAG Chatbot System
# Python 3.9+ required

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
pydantic==2.5.0

# AI & Machine Learning
google-generativeai==0.3.1
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0
chromadb==0.4.18
numpy==1.24.3

# Database
pymongo==4.6.0

# Web Scraping
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
scrapy-playwright

# Natural Language Processing
nltk==3.8.1
pyspellchecker==0.7.2

# Scheduling & Background Tasks
schedule==1.2.0
APScheduler==3.10.4

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6

# CORS Support
starlette==0.27.0
